Create pdf term search
This code can be used to scan a folder of pdf files for a list of keywords, create a list of files that DID NOT contain any of those keywords, and create a csv that shows every pdf and which keywords were present in it.

IMPROVEMENT NEEDED: Many of the earlier pew files are only in .doc format. If code can be improved to include those as well, that would be just fab.